Refactor request handlers for /api/repos to match conventions 	modified:   src/server/config/routes.js 	renamed:    src/server/request-handlers/users.js -> src/server/request-handlers/repos.js
