+++ src/server/request-handlers/repos.js
@@ -4,6 +4,31 @@ var request = require('request');
 var users = require('../queries/users.js');
 
 module.exports = {
+  handleGet: function (req, res, next) {
+    var githubId = req.cookies.githubId;
+    users.getOneAsync(githubId)
+      .then(function (userObject) {
+        var options = {
+          url: "https://api.github.com/user/repos",
+          headers: {
+            'User-Agent': 'GitSpy',
+            authorization: 'token '+ userObject.github_token,
+            'content-type': 'application/json'
+          }
+        };
+        request.get(options, function(error, response, body) {
+          if (error){
+            throw new Error(error);
+          } else {
+            res.json(response);
+          }
+        });
+      })
+      .catch(function (err) {
+        console.error(err);
+        res.sendStatus(400);
+      });
+  },
   handlePost : function(req, res, next) {
     var gitID = req.cookies.githubId;
     users.getOneAsync(gitID)
